Minor Changes
I reverted to Graphics (instead of Graphics2D) and finished its
rotation. It has a problem with a loss of precision within the move
method occuring from adding the trig values, which are less then one
and then converting them to ints. This is best seen when trying to move
in a circle.